requirements.txt
openai-agents==0.2.1
groq==0.30.0
firecrawl-py==2.16.1
gradio==5.38.0